pyproject.toml
[project]
name = "products"
version = "0.10.0"
description = "Сервис с продуктовой логикой."
authors = [{name="Дмитрий Панин", email="dimitri-05@mail.ru"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic-postgresql-enum>=1.7.0",
    "asyncpg==0.*",
    "general-settings==0.*",
    "litestar-granian>=0.13.0",
    "litestar-utils>=0.1.5",
    "litestar[jwt,redis,sqlalchemy,structlog]>=2.16.0",
    "uuid-utils==0.*",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "commitizen==4.*",
    "mypy==1.*",
    "pytest==8.*",
    "pytest-asyncio==1.*",
    "pytest-cov>=7.0.0",
    "ruff==0.*",
]

[project.urls]
Repository = "https://gitlab.com/pochini.online/backend/user-id"

[[tool.uv.index]]
name = "internal-proxy"
url = "https://gitlab.com/api/v4/groups/pochini.online/-/packages/pypi/simple"
publish-url = "https://gitlab.com/api/v4/projects/70263764/packages/pypi"  # Setting -> General -> Project ID
default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.commitizen]
tag_format = "$version"
version_provider = "uv"
bump_message = "сi: новая версия $new_version"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "Q",  # flake8-quotes
    "EM",  # flake8-errmsg
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "D",  # pydocstyle
    "ANN",  # flake8-annotations
    "T20",  # flake8-print
    "PT",  # flake8-pytest-style
    "SIM",  # flake8-simplify
    "ARG",  # flake8-unused-arguments
    "RUF",  # ruff-specific rules
    "TD",  # flake8-todos
    "ASYNC",  # flake8-async
    "S",  # flake8-bandit
    "BLE",  # flake8-blind-except
    "COM",  # flake8-commas
    "PIE",  # flake8-pie
    "RSE",  # flake8-raise
    "TID252",  # flake8-tidy-imports.relative-imports
    "TC",  # flake8-type-checking
    "N",  # pep8-naming
    "PLR",  # Pylint Refactor
    "FURB",  # refurb
    "TRY",  # tryceratops
    "PERF",  # perflint
]

ignore = [
    "COM812",
    "RUF001",
    "RUF002",
    "RUF003",
    "RUF012",
    "S104",
    "D107",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]
"migrations/**" = ["D", "TD"]

[tool.ruff.lint.pylint]
max-args = 5
max-statements = 25

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
multiline-quotes = "double"
docstring-quotes = "double"

[tool.mypy]
enable_error_code = ["deprecated"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.alembic]
prepend_sys_path = "src:."
script_location = "migrations"
file_template = "%%(year)d%%(month).2d_%%(day).2d_%%(slug)s_%%(rev)s"
timezone = "Europe/Moscow"
truncate_slug_length = "40"
version_path_separator = "os"
output_encoding = "utf-8"

[[tool.alembic.post_write_hooks]]
name = "ruff"
type = "exec"
executable = "%(here)s/.venv/bin/ruff"
options = "format REVISION_SCRIPT_FILENAME"

[[tool.alembic.post_write_hooks]]
name = "ruff"
type = "exec"
executable = "%(here)s/.venv/bin/ruff"
options = "check --fix REVISION_SCRIPT_FILENAME"
